requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-ollama
chromadb
pdfplumber
httpx
pytest
google-generativeai
python-multipart
langchain-community
langchain-google-genai